Resolve more MSVC warnings (#2549)

+++ vowpalwabbit/parse_regressor.cc
@@ -350,7 +350,8 @@ void save_load_header(
 
       // TODO: validate ngram_len?
       auto* g_transformer = all.skip_gram_transformer.get();
-      uint32_t ngram_len = (g_transformer != nullptr) ? g_transformer->get_initial_ngram_definitions().size() : 0;
+      uint32_t ngram_len =
+          (g_transformer != nullptr) ? static_cast<uint32_t>(g_transformer->get_initial_ngram_definitions().size()) : 0;
       msg << ngram_len << " ngram:";
       bytes_read_write +=
           bin_text_read_write_fixed_validated(model_file, (char*)&ngram_len, sizeof(ngram_len), "", read, msg, text);
@@ -379,7 +380,8 @@ void save_load_header(
       bytes_read_write += bin_text_read_write_fixed_validated(model_file, nullptr, 0, "", read, msg, text);
 
       // TODO: validate skips?
-      uint32_t skip_len = (g_transformer != nullptr) ? g_transformer->get_initial_skip_definitions().size() : 0;
+      uint32_t skip_len =
+          (g_transformer != nullptr) ? static_cast<uint32_t>(g_transformer->get_initial_skip_definitions().size()) : 0;
       msg << skip_len << " skip:";
       bytes_read_write +=
           bin_text_read_write_fixed_validated(model_file, (char*)&skip_len, sizeof(skip_len), "", read, msg, text);
